refactor: phase 2 shared contracts (chat, listen, songs)
