+++ shared/index.ts
@@ -1,1 +1,4 @@
 export * from "./contracts/ai-models";
+export * from "./contracts/chat-rooms";
+export * from "./contracts/listen";
+export * from "./contracts/songs";
