+++ src/api/rooms.ts
@@ -1,40 +1,20 @@
 import { apiRequest } from "@/api/core";
+import type {
+  BulkMessagesResult,
+  CreateRoomData,
+  Message,
+  Room,
+} from "@ryos/shared/contracts/chat-rooms";
 
-export interface RoomSummary {
-  id: string;
-  name: string;
-  type?: "public" | "private" | "irc";
-  createdAt: number;
-  members?: string[];
-  userCount: number;
-  ircHost?: string;
-  ircPort?: number;
-  ircTls?: boolean;
-  ircChannel?: string;
-  ircServerLabel?: string;
-}
+export type RoomSummary = Room;
 
-export interface RoomMessage {
-  id: string;
-  roomId: string;
-  username: string;
-  content: string;
-  timestamp: number;
+export type RoomMessage = Message & {
   clientId?: string;
-}
+};
 
-export interface CreateRoomPayload {
-  type: "public" | "private" | "irc";
-  name?: string;
-  members?: string[];
-  /** Registered server id from GET /api/irc/servers (required for non-admin IRC rooms). */
-  ircServerId?: string;
-  ircHost?: string;
-  ircPort?: number;
-  ircTls?: boolean;
-  ircChannel?: string;
-  ircServerLabel?: string;
-}
+export type CreateRoomPayload = CreateRoomData & {
+  type: NonNullable<CreateRoomData["type"]>;
+};
 
 export async function listRooms(): Promise<{ rooms: RoomSummary[] }> {
   return apiRequest<{ rooms: RoomSummary[] }>({
@@ -54,16 +34,10 @@ export async function getRoomMessages(
 
 export async function getBulkMessages(
   roomIds: string[],
-): Promise<{
-  messagesMap: Record<string, RoomMessage[]>;
-  validRoomIds: string[];
-  invalidRoomIds: string[];
-}> {
-  return apiRequest<{
-    messagesMap: Record<string, RoomMessage[]>;
-    validRoomIds: string[];
-    invalidRoomIds: string[];
-  }>({
+): Promise<BulkMessagesResult & { messagesMap: Record<string, RoomMessage[]> }> {
+  return apiRequest<
+    BulkMessagesResult & { messagesMap: Record<string, RoomMessage[]> }
+  >({
     path: "/api/messages/bulk",
     method: "GET",
     query: { roomIds: roomIds.join(",") },
